Changes link for syslinux_url in script to stop it hanging.
Although the old link did work and download the file when put into a web browser, it did not work in the script for some reason. When the script got to this step, it simply hang and didn't download the file or continue. The new link now works in the script.